Merge branch 'l3mdev-send-enslaved'
David Ahern says:

====================
net: l3mdev: Allow send on enslaved interface

First patch preps for the second. The second is required for several use
cases such as ping on an interface and BFD that need to send packets on
a specific interface, including ones enslaved to a VRF device.

v2
- fixed brackets on both patches per comment from DaveM
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 20e61335d8db0caf0c50236c1c9c2ea546f61f1a ]