Merge branch 'netdev_lockdep_set_classes'
Eric Dumazet says:

====================
net: better lockdep annotations

Introduction of qdisc->running seqcount added lockdep false positives.

While chasing the bug, it came to me that we had a lot of copies of the
same stuff in virtual drivers.

This patch series has the qdisc->running fix (considers that a trylock
is attempted in lockdep terminology), and adds a generic helper so
that we no longer have to patch many virtual drivers when a new per-device
or per-qdisc lock is added.

Thanks to David Ahern for reporting the issue and testing my patches :)
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: cf515802043cccecfe9ab75065f8fc71e6ec9bab ]