ARM: dts: sun8i-q8-common: Add support for USB wifi controllers
Some of the sun8i q8 boards have an usb wifi controller, on other
variants this will result in an used usb root-hub, but the best
way to deal with wifi on this boards is to simply let the kernel
auto-detect usb or sdio wifi controllers.

This has been tested on an a23 based q8 tablet with a RTL8188ETV wifi
controller.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: 1cdc0aa5aaf995a571550cbf8d57ae9e1454a141 ]