dt-bindings: usb: Fix Tegra XUSB example
The DT bindings example for the Tegra XUSB controller omits the 'lanes'
subnode in the XUSB pad controller which is required according to the DT
bindings documentation for the Tegra XUSB pad controller[0]. In addition
to this the phy-names with the suffix 'utmi' are also not valid and
should have the suffix 'usb2'. Correct both the XUSB pad controller pad
path and phy-names for the XUSB example.

[0]: Documentation/devicetree/bindings/phy/nvidia,tegra124-xusb-padctl.txt

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: b8c5c990a0fc6b00b7f7256864342e3afc24a392 ]